Move handles as well
